OQ3Semantics

Documentation for OQ3Semantics.

OQ3Semantics.Angle

OQ3Semantics.Angle

Base.eps

OQ3Semantics.bitsizeof

OQ3Semantics.Angle — Type

Angle{T<:Unsigned}

NOTE: this doc string is incorrect for _SCALE = 2.0 * pi. Represents an angle in "turns". A turn is a real number in [0, 1). This struct represents real values from 0 to 1 - 1/2^n, separated by increments of 1/2^n where n is the width of T in bits.

OQ3Semantics.Angle — Method

Angle{T}(x::AbstractFloat) where T <: Unsigned

Construct an Angle from a floating point number x in [0, 1). The returned value represents the closest representable angle in turns.

Base.eps — Method

eps(a::Angle)
eps(::Type{<:Angle})

The increment between consecutive representable values of typeof(a) in turns.

OQ3Semantics.bitsizeof — Method

bitsizeof(x)

Return the size of x in bits. This is 8 * sizeof(x).